Group buttons together and add eventlisteners
Eventlisteners were added to the num buttons such that when the user
presses the num keys, the numbers are stored in the display. Once the
user presses an operator key, the displayValue is stored into the first
num and the automatically clears once the user begins pressing in the
second number. The result is displayed when the user presses the equal
button. Additionally, the clear button resets everything to how it was
when the calculator started.

Work still needs to be done for when the user continously calculates
multiple numbers using multiple operators (ex. 2 * 2 + 1).