Reorganize eventlisteners and related functions
New functions were created to be called when different situations call
for the same result. Eventlisteners were modified to behave differently
according to each specific situation (what the previous button pressed
was, whether the firstNum is undefined, whether the secondNum is
undefined... etc).

Edge cases still need to be considered. The calculator does not work
correctly if the user presses the buttons in the wrong orders.