Add css to calculator to make more accessible
Basic css elements were added to make the calculator easier to view and
use. Font and button sizes were increased to make the calculator easier
to use. Additionally, the display alignment was changed so that when
the user enters a number, the numbers are displayed starting from the
right-hand side of the display screen.

Javascript needs to be added so that decimal results do not stretch out
the display screen and visually mess up the calculator. Additionally
need to fix edge case where user presses the equal sign without pressing
an operator button.