Fix major calculation bug with equal starting
A major bug was that when the user pressed the equal sign consecutively,
instead of reusing the second number for the next calculation, the
second number would be set to the first number. This was fixed by
keeping track of the previous button pressed by the user and using if
statements.

The calculator still needs to deal with division by 0 as well as
keeping decimal answers within the digit limit of the calculator.