Large and small number edge cases plus bug fixes
If a number's digit exceeds 13 digits, the number is expressed in
scientific notation so as not to overflow the display screen.

Additionally, division by 0 leads to an ERROR display and the user
must either reenter a number or press the clear button for calculations
to resume.

Bugs associated with pressing the operator button and equal sign before
a number was entered was fixed by using conditional expressions.