Add keyboard functionality to calculator
Event handlers were rearranged into independent functions so that they
could be called for "click" events as well as "keyup" events.
Inside these functions, conditionals were used to distinguish between
the two types of events and to .

The delete button still needs to be activated and as well as the decimal
button.