Add delete button so user can undo their input
Delete button was enabled so that when the user inputs a number, they
are able to backspace in case they mistakenly entered a number. The
keyboard backspace was also enabled.

Users still need to be able to input decimal numbers, and the
calculations need to be changed from integer math to floating number
math.