Fixed decimal bug using conditional statement
Previous decimal placement inconsistency was fixed by adding the
num-button class to the decimal. This however, made it more difficult
to prevent the user from pressing the decimal button more than once
for a single number value. To combat this, a conditional statement was
used to check if the num-button pressed/clicked was a decimal and if
the decimal had already been clicked/pressed.